requirements.txt
Flask>=2.2.2
Pillow>=9.4.0
rembg>=2.0.32
numpy>=1.22.0
onnxruntime>=1.12.1
scikit-image>=0.19.3
aiohttp>=3.8.1
gunicorn>=20.0.4
